Handouts-Textbook-Micro and Nanofabrication (MEMS)-Update
Learn the fundamentals of microfabrication and nanofabrication by using the most effective techniques in a cleanroom environment….Microfabrication and nanofabrication are the basis of manufacturing for nearly all modern miniaturized systems that are ubiquitously used in our daily life. Examples include; computer Chips and integrated sensors for monitoring our environment, cars, mobile phones, medical devices And more. Micro--‐ and nanofabrication can be taught to students and professionals by textbooks and ex--‐ cathedra lectures, but the real learning comes from seeing the manufacturing steps as they happen.